requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymodbus==3.5.2
redis==5.0.1              # Use redis instead of aioredis for Python 3.11 compatibility
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pyyaml==6.0.1
prometheus-client==0.19.0
structlog==23.2.0
